Rework where FluentValidation runs
Now runs when invoking IRequestHandlers, which gives more flexibility in API binding (E.g. a UserId might not be set until the endpoint handler sets it before calling the request handler).

Also aligns with being able to use the base `ProcureHub` project in different hosts, and not depend on running FluentValidation with API binding.